# Fixes #1081: UglifyJS2 should remove unused and dead_code